Flask app build, HTML, CSS, workflow yml and render yaml created.
