+++ render.yaml
@@ -1,0 +1,7 @@
+services:
+  - type: web
+    name: slot-machine
+    env: python
+    plan: free
+    buildCommand: pip install -r requirements.txt
+    startCommand: python slots.py
