+++ requirements.txt
@@ -15,6 +15,7 @@ cycler==0.12.1
 distlib==0.3.8
 dulwich==0.21.7
 ebcdic==1.1.1
+flask==3.0.3
 fastjsonschema==2.19.1
 fernet==1.0.1
 filelock==3.13.1
