+++ static/styles.css
@@ -1,0 +1,40 @@
+body {
+    font-family: Arial, sans-serif;
+    text-align: center;
+    background-color: #f0f0f0;
+    color: #333;
+}
+
+h1 {
+    margin-top: 20px;
+}
+
+form {
+    margin-bottom: 20px;
+}
+
+label {
+    margin-right: 10px;
+}
+
+button {
+    margin: 10px;
+    padding: 10px 20px;
+    font-size: 16px;
+    cursor: pointer;
+}
+
+#slot-machine {
+    display: flex;
+    justify-content: center;
+    margin-top: 20px;
+}
+
+.symbol {
+    font-size: 36px;
+    margin: 0 10px;
+    padding: 10px;
+    background-color: yellow;
+    border: 2px solid #333;
+    border-radius: 5px;
+}
