+++ templates/index.html
@@ -1,0 +1,144 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Slot Machine</title>
+    <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='styles.css') }}">
+</head>
+<body>
+    <h1>Slot Machine Game</h1>
+    <form id="set-budget-bet-form">
+        <label for="budget">Budget:</label>
+        <input type="number" id="budget" name="budget" required>
+        <label for="bet">Bet:</label>
+        <input type="number" id="bet" name="bet" required>
+        <button type="submit">Set Budget & Bet</button>
+    </form>
+    <button id="spin-button">Spin</button>
+    <button id="auto-spin-button">Auto Spin</button>
+    <button id="cancel-auto-spin-button" style="display:none;">Cancel Auto Spin</button>
+    <div id="slot-machine">
+        <!-- Slot machine columns will be displayed here -->
+    </div>
+    <div id="message"></div>
+    <div id="budget-display">Budget: 0.0</div>
+    <div id="spin-counter-display">Spin Count: 0</div>
+    <div id="auto-spin-count-display">Auto Spin Count: 0</div>
+
+    <script>
+        document.getElementById('set-budget-bet-form').onsubmit = function(event) {
+            event.preventDefault();
+            const budget = document.getElementById('budget').value;
+            const bet = document.getElementById('bet').value;
+            fetch('/set_budget_bet', {
+                method: 'POST',
+                headers: {'Content-Type': 'application/json'},
+                body: JSON.stringify({budget, bet})
+            })
+            .then(response => response.json())
+            .then(data => {
+                if (data.success) {
+                    document.getElementById('budget-display').innerText = `Budget: ${budget}`;
+                    document.getElementById('spin-counter-display').innerText = `Spin Count: 0`;
+                }
+            });
+        };
+
+        document.getElementById('spin-button').onclick = function() {
+            spin();
+        };
+
+        document.getElementById('auto-spin-button').onclick = function() {
+            const count = prompt("Enter number of auto spins:");
+            if (count > 0) {
+                fetch('/auto_spin', {
+                    method: 'POST',
+                    headers: {'Content-Type': 'application/json'},
+                    body: JSON.stringify({count})
+                })
+                .then(response => response.json())
+                .then(data => {
+                    if (data.success) {
+                        document.getElementById('auto-spin-count-display').innerText = `Auto Spin Count: ${data.auto_spin_count}`;
+                        auto_spin();
+                    }
+                });
+            }
+        };
+
+        document.getElementById('cancel-auto-spin-button').onclick = function() {
+            fetch('/cancel_auto_spin', {method: 'POST'})
+            .then(response => response.json())
+            .then(data => {
+                if (data.success) {
+                    document.getElementById('auto-spin-count-display').innerText = `Auto Spin Count: 0`;
+                }
+            });
+        };
+
+        function spin() {
+            fetch('/spin', {method: 'POST'})
+            .then(response => response.json())
+            .then(data => {
+                if (data.success) {
+                    document.getElementById('budget-display').innerText = `Budget: ${data.budget}`;
+                    document.getElementById('spin-counter-display').innerText = `Spin Count: ${data.spin_counter}`;
+                    const slotMachine = document.getElementById('slot-machine');
+                    slotMachine.innerHTML = '';
+                    data.columns.forEach(col => {
+                        const colDiv = document.createElement('div');
+                        col.forEach(symbol => {
+                            const symbolDiv = document.createElement('div');
+                            symbolDiv.classList.add('symbol');
+                            symbolDiv.innerText = symbol;
+                            colDiv.appendChild(symbolDiv);
+                        });
+                        slotMachine.appendChild(colDiv);
+                    });
+                    if (data.auto_spin_active) {
+                        document.getElementById('cancel-auto-spin-button').style.display = 'block';
+                        document.getElementById('auto-spin-button').style.display = 'none';
+                    } else {
+                        document.getElementById('cancel-auto-spin-button').style.display = 'none';
+                        document.getElementById('auto-spin-button').style.display = 'block';
+                    }
+                } else {
+                    document.getElementById('message').innerText = data.message;
+                }
+            });
+        }
+
+        function auto_spin() {
+            fetch('/spin', {method: 'POST'})
+            .then(response => response.json())
+            .then(data => {
+                if (data.success) {
+                    document.getElementById('budget-display').innerText = `Budget: ${data.budget}`;
+                    document.getElementById('spin-counter-display').innerText = `Spin Count: ${data.spin_counter}`;
+                    document.getElementById('auto-spin-count-display').innerText = `Auto Spin Count: ${data.auto_spin_active ? data.auto_spin_count : 0}`;
+                    const slotMachine = document.getElementById('slot-machine');
+                    slotMachine.innerHTML = '';
+                    data.columns.forEach(col => {
+                        const colDiv = document.createElement('div');
+                        col.forEach(symbol => {
+                            const symbolDiv = document.createElement('div');
+                            symbolDiv.classList.add('symbol');
+                            symbolDiv.innerText = symbol;
+                            colDiv.appendChild(symbolDiv);
+                        });
+                        slotMachine.appendChild(colDiv);
+                    });
+                    if (data.auto_spin_active) {
+                        setTimeout(auto_spin, 500);
+                    } else {
+                        document.getElementById('cancel-auto-spin-button').style.display = 'none';
+                        document.getElementById('auto-spin-button').style.display = 'block';
+                    }
+                } else {
+                    document.getElementById('message').innerText = data.message;
+                }
+            });
+        }
+    </script>
+</body>
+</html>
